Fix paths, extract FAUNA_READ_KEY
