+++ packages/web/script/fauna_client.ts
@@ -1,15 +1,14 @@
-// import * as dotenv from 'dotenv';
+import * as dotenv from 'dotenv';
 import faunadb from 'faunadb';
 
-// dotenv.config({ path: __dirname + '/../../../.env' });
-export const FAUNA_KEY = process.env.FAUNA_KEY ?? '';
-export const FAUNA_API_KEY = process.env.FAUNA_API_KEY ?? '';
+dotenv.config({ path: __dirname + '/../../../.env' });
+export const FAUNA_WRITE_KEY = process.env.FAUNA_WRITE_KEY ?? '';
 export const FAUNA_API_URL = 'https://graphql.us.fauna.com/graphql';
 
 // https://docs.fauna.com/fauna/current/drivers/javascript
 // https://fauna.github.io/faunadb-js/
 export const client = new faunadb.Client({
-  secret: FAUNA_KEY,
+  secret: FAUNA_WRITE_KEY,
   domain: 'db.us.fauna.com',
   port: 443,
   scheme: 'https',
