+++ packages/web/src/fauna/indexes/by_highest_last_sale_index.ts
@@ -1,4 +1,4 @@
-import { client, q } from '../../fauna_client';
+import { client, q } from '../../../script/fauna_client';
 
 // Index of highest last Eth sale on OpenSea
 //
