+++ packages/web/src/fauna/indexes/by_most_affordable_index.ts
@@ -1,4 +1,4 @@
-import { client, q } from '../../fauna_client';
+import { client, q } from '../../../script/fauna_client';
 
 const sourceObj = {
   collection: q.Collection('DopeToken'),
