+++ packages/web/src/fauna/indexes/dope_text_search_index.ts
@@ -1,4 +1,4 @@
-import { client, q } from '../../fauna_client';
+import { client, q } from '../../../script/fauna_client';
 
 // Concatenates all strings, then splits using NGram feature.
 // Produces a bound field array of tokens we can then use
