+++ packages/web/src/fauna/indexes/on_sale_index.ts
@@ -1,4 +1,4 @@
-import { client, q } from '../../fauna_client';
+import { client, q } from '../../../script/fauna_client';
 
 // Indexes if On Sale with OpenSea
 // https://docs.fauna.com/fauna/current/tutorials/indexes/search
